Removed audio (please see details for explanation).
Unfortunately, it would appear that the current version of SimpleAudioEngine causes a crash in OpenGL when the app loses focus, so the small amount of audio I am using has to be removed. Otherwise, much more important features like "Get Directions", "View Website" and "Get Tickets" will cause the app to crash.